requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv

langchain
langchain-community
langchain-google-genai

pypdf
faiss-cpu